src/pyspark/requirements.txt
google-cloud-bigquery==3.13.0
pyspark==3.2.0